+++ Sources/EZNetworking/Error/NetworkingError.swift
@@ -7,7 +7,7 @@ public enum NetworkingError: Error {
     // HTTP Status Code errors
     case httpRedirectError(HTTPRedirectionStatus)   /// 3xx status code errors
     case httpClientError(HTTPClientErrorStatus)     /// 4xx status code errors
-    case httpServerError(HTTPServerStatus)          /// 5xx status code errors
+    case httpServerError(HTTPServerErrorStatus)          /// 5xx status code errors
 
     // URL Errors
     case urlError(URLError)                                  /// any URL error
